(GH-1378) allow external log4net config file
If choco finds a log file next to it by the name of `log4net.config.xml`
(a little bit different on purpose), it will use that file instead of
the built-in configuration for log4net. Use this with caution as it
could cause certain logging items to disappear when the external config
does not include the necessary logging sections. It may even cause
choco to error.