(GH-1383) Fix - no pkg args cause "value cannot be null"
When there are packages that do not have a saved .arguments file, if
someone has useRememberedArgumentsForUpgrades turned on, it can cause
"Value cannot be null" issues when running outdated or upgrade for the
packages that do not have these files. This is seen for packages that
were installed or last upgraded before Chocolatey v0.10.4.