(GH-1133) Package dependency check globally available
The check to determine if a package is considered a dependency should
be a globally available check. This allows other code to take advantage
of checking to see if a package is considered a dependency or not
during install, upgrade, or uninstallation. Move that out to a static
class so that the check can be called with other functions.