(GH-1396) API Fix: config output is explicitly set
When running commands with GetChocolatey, it appears the output is
getting explicitly set to RegularOutput. Change this so that it is not
set at all here and managed by the user when they run the command. If
a call wants to be silent, with this here it would not be allowed.
Allow that call to be silent.