(GH-1400) API - Ensure one instance of GetChocolatey
Surround `Lets.GetChocolatey` with a mutex to ensure only one instance
of `GetChocolatey` is created. There was an issue when accessing the
Chocolatey API, in that you could have multiple threads creating an
instance of `GetChocolatey` at the same time and confusing some of the
setup, causing things to fail. Instead of multiple instances, there
should be a singular instance of `GetChocolatey` .