(GH-1459) Fix: Get-PackageParameters handles urls
Previously, when handling parameters with url values, which have both
`:` and `/`, Get-PackageParameters would split on those parameters.
This made it near impossible to pass through urls without finessing the
values.

Switch to using a regular expression that can handle the complexity of
parameters:

* Split on start of string and `/` or a space and then `/`.
* The splitter for each item name and value is either `:` or `=`.
* A flag should be set true if present (an item without a value).
* An item value can contain spaces colons and equals signs, pretty much
anything except for a space followed by "/".