(GH-1490) Remove quotes/apostrophes surrounding parameter
It was noted that there is a discrepancy with how the built-in
`Get-PackageParameters` behaves differently and in an incompatible way
as compared to the the core extension's old version of
`Get-PackageParameters`. The built-in Get-PackageParameters should do at
least what core extension does and then enhance on that with more.

To remove the noted incompatibility, it should handle quotes and
apostrophes surrounding the items in the same way, which is to remove
them as it loads the parameter values into the hash table.