(GH-1397) Reduce traffic - choco queries
Ensure all Chocolatey queries are sped up by using better selection on
the front part of the call to ensure what is returned is the result that
is being requested without a huge amount of traffic to get to that value.

Previously, a call to find a package or even a version for upgrade might
send queries to remote sources to provide all of the versions of a package
starting with the oldest version to the latest. This results in terrible
performance, especially as a source continues to have more and more versions
of a package, the performance gets worse. Instead, for install/upgrade/outdated
scenarios where choco is querying service-based url, it should make one call
ensuring it receives the version it is looking for.