(GH-1517) Fix: Given path's format is not supported
When setting output directory, if you do it in the proper manner as
suggested by the documentation, it will return an error with "The given
path's format is not supported." This is because the option is not
removing the surrounding quotes like the other options do.

Allow passing `--outputdirectory="'c:\somepath'"` in the proper manner
so that it matches consistency with all of the other options.