(GH-1476) Ensure docs are for Get-PackageParameters
When renaming the function to force it's precedence over
Get-PackageParameters in the core extension, that causes the documents
to generate the name based on the renamed version. That needs to use
the original name. Change it back in generate docs just after importing
the module but prior to generating documentation.