(GH-1524) Capture password when only user is provided
Normally, when full credentials have not been provided, there will be a
prompt from the credential provider to ask for that user name and
password. However if something is routing through the background
service, that is too late in the process to provide that information as
it would happen during a non-interactive part of the session - thus it
needs to be asked for early on before the background service is called.