(GH-1529) Fix: install of choco exit code 1
With Chocolatey v0.10.9 and #1401, we added a call to choco to
initialize the config file during installation so that it was generated
during and administrative run. That was done in commit 35add3b7b.

A call to choco with no switches results in an exit code of 1 instead
of 0. Recently a call to choco to create the config file on install was
added as the last step in installing Chocolatey. Unfortunately this
resulted in an install or upgrade of Chocolatey resulting in a 1
instead of a 0. Call choco -v instead ot show the version, which exits
with a 0 and also ensures the first time generation of the
chocolatey.config file.