(GH-991) Install - Modify profile if file exists & empty
When installing Chocolatey, install the tab completion if the file
exists and is empty. Previously, it would check authenticode signature
on an empty file if it existed and that would error, assuming that the
file was authenticode signed. Now check the file size first to be sure
that it is bigger than 4 bytes. If it is not, then it an empty file and
it is safe to assume it is not signed either.