(GH-1540) Fix - AutoUninstaller can't find escaped quotes
The change made for GH-1505 caused the registry information to escape a
little too well, so some of the quotes and apostrophes were being saved
as `&apos;` and `&quot;`. When it came time for automatic uninstaller
to remove the software that was found in the snapshot, it didn't know
how to work with those escaped double quotes and gave up very quickly.

To fix this, don't escape those two items, but handle any that may have
already had this adjustment.