(GH-2092) Take all registry keys into account
It is possible that the installation of an application can generate
multiple registry keys within the snapshot file that is craeted.  Rather
than looking at only the first registry key when deciding if an installed
application is covered by a Chocolatey package, look at all keys within
the snapshot file.