(GH-1623) Use PSVersionTable for Version
When determining the PowerShell version, using `$host.Version` is ill-advised -
it is dependent upon the host and may not be an accurate value. In the case of
Visual Studio Code, it was coming up with the version of VS Code and not
PowerShell's version. This leads to unexpected behavior. Instead, to make a
determination of the version of PowerShell, use `$PSVersionTable.PSVersion`.

See this [link](https://blogs.technet.microsoft.com/heyscriptingguy/2014/10/13/powertip-check-version-of-powershell/) for more information.