(GH-1644) Note registry uninstall may need split
For uninstallation of software when pulling values from the uninstaller
registry, the value pulled for the uninstall string is likely to be a
full value that needs split up into the file to run and the arguments
to pass to it. Provide documentation notes to give a frame of
reference, not every installer is going to be the same so don't
approach putting a more generic way of approaching this at this time,
just documentation tidbits for package maintainers to determine the
best method to move forward with.