(GH-1397) Fix: Ensure config reset
Ensure the config is reset to a known point with each package being run
in an operation. Previously it was thought this was happening, but the
way the reset occurred was actually changing the original configuration
and the current. Ensure that is getting reset accordingly by doing the
deep copy to set config each time through.