(GH-1292) add note on manifest caching
Windows behavior of manifest caching is enough to make even the best of
admins and developers want to tear out their hair until they realize
there is a cache and changes to the manifest are completely ignored by
Windows. So leave a breadcrumb here so that folks know if they make a
change to the manifest file, they will need to "touch" the file they
are trying to get the manifest to take effect for - in this case
choco.exe. The modification date will need changed to clue
Windows in that there is a new manifest.

This is only when changing the file AFTER the exe / modification date
has been run even once. It's better to just say to always update the
modification date as someone may not know whether the file has been ran
before.