(GH-1503) Warn on incorrectly named license file
If choco does not find a valid license, but it finds other files in the
license folder, assume that a user misnamed the file or did not rename
it properly. Log an error and warning with information on how to
correct the issue.

If a user puts the license in the top level folder, and also if they
forget to rename the file, log an error and warning with information on
how to correct the issue.