(GH-1735) Globally resolve assemblies w/out locking
When dynamically loading assemblies, do not lock those assembly files
on the file system. Instead of loading the assembly file directly, read
the Assembly binary into memory and then load the result. As .NET will
happily load an assembly multiple times, synchronize and ensure that if
the assembly is already loaded, use that instead of attempting to load
the assembly again.