(GH-1646) upgrade - skip non-installed packages
Previously, if a package was not installed, during upgrade it would
install the package to meet the end state of having a package installed
and up to the latest version. However in some cases, it may be desired
to have the installation skip those packages and continue on, passing a
successful exit code instead of a failure like it would with
`--fail-on-not-installed` (which has similar behavior).

Provide a feature switch `skipPackageUpgradesWhenNotInstalled` that can
be overridden in the upgrade command with either
`--skip-if-not-installed` or `--install-if-not-installed` (the
later only being necessary when the feature switch has been flipped on).