(GH-1746) Add concept of global validation
This is implemented as a validation engine. Any new validations will
need to implement the IValidation interface, and be added to the
container in the ContainerBinding class.

When running with limited output, error messages will still be shown to
the user, but warning messages will be suppressed.

Initial validation is provided which errors if attempting to halt when
a reboot is detected, but when usage of package exit codes is turned
off.  This is an error, since without package exit codes turned on,
Chocolatey is not able to detect what actually happened when a package
is installed/upgraded/uninstalled.