(doc) move quoting options point around
One of the bullet points mentions that "the next section applies",
however there was a bullet point that was later inserted about periods
in PowerShell. That was inserted just after the aforementioned bullet
point, invalidating or confusing folks on what the next section was.
Instead move this point about periods below the next point to keep the
other two together.