(GH-1747) Fix: passing timeout switch of 0 is ignored
Passing the execution timeout of 0 is ignored and the default value is
used. It appears this situation was not explicitly being looked for as
zero is typically the return value when parsing and not getting a value
back. Look specifically if the user-provided string was zero and use
that to set infinite runtime.