(maint)(log) update message about number of tries
Previously the message coming across made it sound like the number of
attempts was the number of retries after the first time failed. This is
not the case - the number passed in is the number of attempts it will
do and that needs to be at least 1 to jive with the messaging. Fix the
wording to state "tries" instead of "retries".