(GH-1758) control enhanced exit codes w/feature
In #1602 and #1724, enhanced exit codes were added to provide more
intentional exit codes that determine the state of what happened during
the run. This would allow simply seeing a 0 from outdated and knowing
that all packages are up to date, or seeing a 2 and knowing that
packages need updated. This allows for better scripting based on exit
codes.

However, there are some existing integrations that might be broken on
taking on a newer version of Chocolatey if the enhanced exit codes are
not being looked for yet. To allow compatibility to older systems, add
a feature switch that can be disabled to provide the older behavior of
0 or 1 exit codes.