(GH-1771) Download nupkg, extract nuspec
Previously, when Chocolatey was downloading a nupkg, it would rely on
the metadata information from an OData feed to be returned properly.
Unfortunately, not all packaging repositories that support OData
repositories return this data in a proper manner. Due to this aspect,
download the nupkg first and extract the nuspec from that file instead
or using the metedata returned from the `DataServicePackage` (which
is simply an XML stream from a URL).