(GH-1768) Ignore file rename operations check
In some scenarios, this check was causing Chocolatey to always report
that the system was in a pending reboot state. Even after a reboot
Chocolatey would still report that it was still pending a reboot.  After
discussion, it was determined that this check has a history of causing
problems, and as a result, the decision was made to remove it.  It will
still do the check, which will now only show up in the debug log, but
this check will no longer go into the overall pending reboot check.