(maint) move find_package to NuGetList
Listing or searching using some of NuGet.Core components should be in
a common place - that is NuGetList.cs currently. Move the optimized
method to where it can be used everywhere.

Additionally, add smarts to the method to determine if one is using the
older methods of search when repository optimizations are not available.