(GH-1781) fix: use ProviderPath for Get-Location
When seeing the following error:
`ERROR: Exception calling "Start" with "0" argument(s): "The directory
name is invalid"` this can stem off of permissions issues or an
incorrectly formatted directory being passed. When using Get-Location,
if that location is a UNC path, PowerShell shows that with
`Microsoft.PowerShell.Core\FileSystem::\\server\share`, resulting in
that entire string to be passed as the PATH for the working directory
that is passed to
`System.Diagnostics.Process.StartInfo.WorkingDirectory`. Unfortunately
that doesn't evaluate to an acutal valid path, thus the error is
produced. Instead, the path should be evaluated with ProviderPath,
which removes the `Microsoft.PowerShell.Core\FileSystem::` from the
actual path, resulting in a valid path using a UNC location.

Also make a determination if the path is not set (such as with things
that are run during provisioning) and provide a valid working directory
using cache location.