(GH - 1808) Add ignore-pinned-messages option to upgrade command
Previously the upgrade command would output messages of the upgrade for
a pinned package.  With this new ignore-pinned-messages, there is a
check prior to outputting these messages to see if the package is pinned
and if this ignore message flag is set.  If it is, it will not output
the messages to the console.