(GH-1020) Docker add ChocolateyInstall Env Var
The ChocolateyInstall environment variable is needed to for docker
choco to not output a warning that it cannot find the Chocolatey
lib directory and use a different path for packages depending on where
the command is run from. Add the environment variable to make it available
in the right location.