(GH-1949) Remove unofficial message on limit output
When attempting to test integrations, folks use -r as part of that
(limit output). When using an unofficial build to attempt that testing,
we previously have shown a message no matter what that tells folks they
are on an unofficial build - that needed to be parsed and ignored where
an official build would not have had that. When using limit output, log
only to the file.