(GH-1949)(build) ensure files unpacked w/unofficial
When building the Chocolatey package without the official strong name
key, it would not properly create the nuget package due to running
`choco unpackself -f` without the additional
`--allow-unofficial-build`. Instead it would throw an error about
needing that passed. Having that additional switch is ignored on
official builds.

Add it to the `nugetPrepare.post.step` file as part of the arguments
to ensure that the nuget package has all of same components that the
officially strong named package would have.