(GH-1949) New official release build configuration
This will be used when preparing for an official release of Chocolatey,
to ensure that the correct strong name key is being used, and also that
the necessary checks are being done when loading the Chocolatey Licensed
assembly into the Chocolatey process.  The FORCE_CHOCOLATEY_OFFICIAL_KEY
compilation symbol if used in if/def statements to make this logic
happen.

A new build.official.bat file was also created to make the calling of
this specific build easier to do as well.