(GH-1949) switch assembly location on official key
Previously, Debug vs Release was the switch to where Chocolatey resolved
its location along with where the licensed dll would be. Now use the
official key setting to determine that switching. This should allow
locating items next to chocolatey if local as well as when installed
at the machine level.

For release official mode, maintain what used to be Release to allow
for that same behavior to continue.