(GH-1949) Use only official/unofficial on resolve
Previously, there was an attempt to use an either or key check when
attempting to find the right assemblies to resolve. Unfortunately that
was a bit optimistic as there is a strong tie in for using one type of
key all the way through to other assemblies.

Only check against one type of key for resolution to allow things to
resolve appropriately. This removes a mix and match thought process
that was optimistic.