(GH-1949) Release mode - use entry assembly to check
When in release mode, a chocolatey.dll could be next to something being
tested. However, most of the time, for testing, it will be located at
the machine level install. When that is the case, the chocolatey
assembly is loaded into memory to keep from locking the files and the
executing assembly path is null, so it uses the current directory.

Instead prefer to look near the entry assembly, then defer to the
environment variable for the machine installation.