(GH-1670) Add noshims options for install and upgrade
The --noshims and --noshimsglobal options allow the user to stop shims
being created for either the package, or the package and its
dependencies.

In addition to this new behaviour, the implementation handles the
creation and removal of shims more robustly, so that shims from other
packages (or the same package) cannot be unintentionally overwritten or
removed.

The new ShimRegistry class keeps track of the currently installed shims
and their corresponding packages. It does this by examining all the exe
files in the shim directory (so old-style .bat and unixy shims are not
supported) and extracting the target file from the binary. If this
target path contains the lib folder, then the package name can be
obtained.

ShimRegistry is updated for each package, just before the Powershell
scripts are run, using `ShimGenerationService::take_snapshot`. When
ShimGenerationService installs any shims for the package, ShimRegistry
provides its existing shim files (ie. those that were not modified or
removed since the last update) so that they can be deleted prior to
installing any new ones.

To stop shims being added from Install-BinFile, a new environment
variable `chocolateyNoShims` is used when running the Powershell
scripts. If a package sets a shim this way, but then forgets to call
Uninstall-BinFile when uninstalling, the shim will be removed anyway if
its target is in the package folder.