(GH-1843) Fix: search exact all versions returns 1
Previously, choco 0.10.15 switched to a new method of finding an exact
package based on the name, however that change also meant it would only
retrieve one version of a package. There is however a scenario of using
the `--exact` switch to retrieve all available versions of a package.
This method was used in some integration tools. The change put in for
0.10.15 broke this scenario completely.

Check for all versions and return all packages for an exact search
along with other search criteria. If not returning all versions,
continue to use the updated method of retrieving a package.