(GH-2055) remove logs/config folder in nupkg
When building choco for official release, Chocolatey uses the machine
installation location for writing out logs and checking for the
existence of a config file. When in a non-official release, it uses
the output folder, which means it will be packaged up inside the
package. This is bad because it later overwrites the existing log
files, and it overwrites the config, resetting/removing any settings
one may be depending on. This only affects a non-official release, or
releases that are used for testing.

Ensure the logs/config folders are removed and not packaged into the
nuget packaging.