(#886) Add properties to packages.config
This commit adds the force and the commandExecutionTimeoutSeconds
attributes to the package element within the packages.config file. These
additional properties will mean that a user will be able to control
these options via the packages.config file when running choco install
packages.config.

Follow up commits will add other attributes to the package element so
that all aspects of the installation can be controlled via the
packages.config file.