(doc) Update script to Statiq format
The docs for Chocolatey are now hosted as a Statiq statically generated
website, and as a result, there has been some changes to how the
generated docs need to "look".  These updates make sure that all the
generated docs are what is now needed, and in addition a "sync" was
performed, as some sections no longer matched what was actually there.