(#2227) Use latest Mono and Ubuntu on Travis
When building the chocolatey code base on Travis, there is a failure
when trying to restore nuget packages. This is due to the fact that an
older version of Mono doesn't contain the newer root certificates
necessary to ensure a valid TLS handshake.  Using a newer version of
Mono ensures that these are in place.

In addition, updating to the latest Ubuntu (bionic).