(#2078) Skip importing Chocolatey GUI Extension
The Chocolatey GUI Extension is a "special" extension in the sense that
it isn't intended to work with Chocolatey directly, but it is shipped as
a "normal" Chocolatey extension. We don't want to load the
chocolateygui.licensed.dll into the choco.exe process, but rather leave
it alone, and allow Chocolatey GUI to load it when required.