(#2379) Correct pack files filter for Linux
If there is not a <files> element in the .nuspec, then the pack command
will gather up a list of all of files the in the directory which is
used to include them in the package. However, the nuspec should not be
included in this list, because it is auto-regenerated. The .nupkg also
should not be included. Therefore, there is a filter to remove the
.nuspec and .nupkg from the list of files.

However, this filter has the directory separator hardcoded as a
backslash. This does not work correctly on platforms where the
directory separator is a forward slash. Therefore, this fixes
the directory separator to be correct depending on what platform choco
is running on.

Co-Authored-By: Kim J. Nordmo <1474648+AdmiringWorm@users.noreply.github.com>