(#2259) Switched to versioned NuGet.Core
Prior to this change, there was no clear indication of what NuGet.Core
assembly was being used within the project, only that it was being used.
With this change in place, we can trace back to the specific package
version that is being used. This package is not pushed to a public
repository currently, but this may change over time.  For now, the
nuget install command was run directly to place this package into the
lib folder.