(#357) Ensure to use full file path
When creating an output export file from a path like "packages.config"
we need to ensure to use the full file path to the file, otherwise it
won't be able to backup the file correctly, if it already exists on the
file system.

Without this, we were seeing errors when trying to ensure that the
directory that the file was contained in was created, since this was an
empty string.