(#2220) Add convert command to allowed list
In a future version of Chocolatey Licensed, a new convert command is
going to be introduced, which will work with nupkg files. We are
preemptively adding the convert command in the next release of
Chocolatey, so that we don't have to release another version nearer the
time we ship the convert command.