(#2287) Improvements to Tab Expansion
Previously, when installing, updating, or uninstalling packages, the
tab expansion would only occur for the first item. Further, the tab
expansion would not work if the search term contained a dash.

Now tab expansion works for all items, and items can contain dashes
as long as it does not start with dash.