(#2304) Fix list --exact -a for prereleases
Due to a prior fix in the logic for `list -e -a`, all packages with the
target ID were being returned, regardless of whether the user uses the
`--pre` flag or not. Since the same search without `--exact` does not
return prerelease versions unless the `--pre` option is also passed, the
behaviour for `list -e -a` should match this.

This fix modifies the search logic used when listing all package
versions with `--exact` to also check for and respect the `--pre` flag
from the configuration context, much like other list commands use the
setting in this file in other places.

Also includes a minor refactor to avoid doing a double query here; now
we only call the Search() function if we're actually going to use those
results. The code paths for AllVersions completely discard that already,
so this just removes the need to call the method unnecessarily.