(#2333) Disable loading of DLL under extensions path
As there isn't a way to determine whether a DLL is a usable PowerShell
module without loading it into memory, we're better off not loading DLLs
directly, as it has caused issues several times in the past
(see: #2078, #1041).
This has previously required us to create an ever-expanding list of
exceptions to the loading behaviour.

Instead, if folks want to load DLLs as powershell modules, they can
package them alongside a PSM1 file that handles the DLL import itself.
This removes the need for Chocolatey to guess which dlls are powershell
modules and which should not be loaded.