(test) Add unit tests for ArgumentsUtility class
This is starting to be used in other places, so we should have some
tests around the execution of the logic contained within this class to
ensure that there are no regressions if/when editing this class.

For now, all the current expectations are covered within the different
TestFixtures, but this should be extended if any changes are made to
the class.