(#2356) Remove .travis.yml file
We are no longer going to be using Travis.  This is for two reasons.
Firstly, there have been some security findings with Travis which
haven't been handled very well:

https://twitter.com/peter_szilagyi/status/1437646118700175360?s=21

And where we were using Travis, i.e. to build on Linux, we are now
doing this with GitHub Actions, so Travis isn't really needed anymore.