(#2364) Rework Linux Dockerfile
Rework the Dockerfile for Linux. Move to a multi-stage build so the
source code is not brought along into the final image (#2364). Add
build args to select the build script and mono version to use. Remove
deprecated MAINTAINER instruction and add metadata with LABEL (#2361).
Add "choco.exe" to path, so both "choco" and "choco.exe" can be used,
like Windows supports (#2362). Change the install location to
"/opt/chocolatey" as that is more in line with the Linux filesystem
hierarchy (#2360). Create the "lib" directory to prevent the warning
about the directory not being found (#2363).

The choco_official_wrapper is used for official builds, where the
"--allow-unofficial" argument is not needed.