(#2374) Remove setting of Authorization property
This property does not actually exist on the HttpWebRequest object:

https://docs.microsoft.com/en-us/dotnet/api/system.net.httpwebrequest?view=netframework-4.0#properties

As such, it can't be used. Based on testing that was completed, the
Authorization header was correctly being set in v0.10.15 of Chocolatey,
and no work was required to be done here.

This commit, essentially puts this file back to how it was in v0.10.15.