(#2383) Use "/" instead of SystemDrive on non-Windows platforms
Linux and MacOS do not normally have a "SystemDrive" environment
variable since they have everything under a single root of "/".
This switches to use "/" instead of "SystemDrive" on non-Windows
platforms. These two instances are the only ones I see in the C#
codebase. There are a couple of uses of "SystemDrive" in the PowerShell
helpers, but those are not run on non-Windows platforms.