(#2377) Add new config value DefaultTemplateName
This will be used to allow the user to control which template should be
used by default. While it is possible to allow the user to override the
default template files, this gives them a little bit more control, to
specify via configuration which template they want to use by default.