(#2377) Use defaultTemplateName if set
If the defaultTemplateName configuration value is set, and the user
hasn't specified a template name at the command line, check to see
whether a template exists with that name. If it does, use this as the
template when generating the new package.

Given that the default value for the new defaultTemplateName
configuration value is an empty string, if the user is directly
overriding the default template files folder, this overriding will still
take place.

Also take into consideration the usage of the --built-in command line
option. If this is set, make sure to respect it. The order of
precedence is the following:

1. choco new --template=name
2. choco new --built-in
3. choco config defaultTemplateName
4. Manually created files in $env:chocolateyinstall\templates\default
5. Fall back onto the built-in template.