(#2382) Update to latest NuGet.Core assembly
This new assembly includes a fix to address an issue when attempting to
pack a Chocolatey nuspec file that contains escaped XML characters on a
Linux machine.  During testing, it was found that a nuspec file that
contains escaped XML characters would work fine on a Windows machine,
but the build would fail on a Linux machine.  A change was applied
upstream to correct this inconsistency.