(maint) Replace unversioned Rhino.Licensing dll
The DLL here is the same DLL used originally, just reinstalled from a
`nuget install` so that we have some traceability as to the version.
The version applied is from a best-guess basis based on a commit in the
chocolatey/rhino-licensing repository on the corresponding date,
although the actual file was created prior. We surmise this was built
locally and used well before the commit was finalised and published.