(#2412) Switch to output warning on failure
This commit updates the chocolateyInstaller module to not output an error
when failing to load the Chocolatey Licensed Extension which would prevent
all helper cmdlets in FOSS being exported.

Instead the importing have been wrapped inside a try/catch block and writes
out a Warning that includes the reason the importing failed.