(#1541) Fix package being added to results if none found
This commit updates the the acquiring of package from prioritized aggregate
repository to not add a package to the result list we iterate through
when a package could not be found.

This change was necessary in case an invalid identifier was used, or
the package exist but is only available on a disabled source.