(#2424) Disable alternate sources on non-windows
This throws an exception if running on a non-windows system for all sources
except for nuget. All of the alternates sources were broken, so this displays a
helpful error message instead of the misc error messages.

Later on, the python and ruby sources could theoretically be fixed to work on
non-windows.