(#2089) Correct variables for package is a dependency check
Two variables appeared to have been swapped accidentally when this
check was created in e91f209. The "package" is the var from the config,
while the "packageName" is the var that is being checked if it is in
the config. This fixes that swap.

This also changes the check for a directory separator in the name to
work on non-windows as well.