(#62) Add unit tests for DefaultPushSource
Adds tests for DefaultPushSource to check it works correctly.

If DefaultPushSource is set, it should set the source assuming that
a source is not passed in explicitly with --source. If a source is
passed in explicitly, then DefaultPushSource should not override it.

If DefaultPushSource is set to "disabled", it should throw an error if
there is not an explicit source passed in. But if an explicit source is
passed in, it should not error.

If the DefaultPushSource is set to the name/alias of a machine source,
it should resolve to that machine source.